+++ R/inflate.R
@@ -16,11 +16,13 @@ regex_example <- paste(regex_example_vec, collapse = "|")
 #' Inflate Rmd to package
 #'
 #' @param pkg Path to package
-#' @param name Name of the resulting vignette
-#' @param rmd Path to Rmarkdown file to inflate
+#' @param flat_file Path to Rmarkdown file to inflate
+#' @param vignette_name Name of the resulting vignette
+#' @param open_vignette Logical. Whether to open vignette file at the end of the process
 #' @param check Logical. Whether to check package after Rmd inflating
 #' @param document Logical. Whether to document your package using \code{\link[attachment:att_amend_desc]{att_amend_desc}}
-#' @param overwrite Logical (TRUE, FALSE) or character ("ask", "yes", "no). Whether to overwrite vignette and functions if already exists.
+#' @param overwrite Logical (TRUE, FALSE) or character ("ask", "yes", "no).
+#' Whether to overwrite vignette and functions if already exists.
 #' @param ... Arguments passed to `rcmdcheck::rcmdcheck()`.
 #'     For example, you can do `inflate(check = TRUE, quiet = TRUE)`, where `quiet` is
 #'     passed to `rcmdcheck::rcmdcheck()`.
@@ -40,7 +42,8 @@ regex_example <- paste(regex_example_vec, collapse = "|")
 #' dev_file <- add_flat_template(template = "full", pkg = dummypackage, overwrite = TRUE)
 #' flat_file <- dev_file[grepl('flat', dev_file)]
 #' fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package"))
-#' inflate(pkg = dummypackage, rmd = flat_file, name = "Exploration of my Data", check = FALSE)
+#' inflate(pkg = dummypackage, flat_file = flat_file,
+#'   vignette_name = "Exploration of my Data", check = FALSE)
 #'
 #' # Explore directory of the package
 #' # browseURL(dummypackage)
@@ -50,9 +53,24 @@ regex_example <- paste(regex_example_vec, collapse = "|")
 #' # pkgdown::build_site(dummypackage)
 #' # Delete dummy package
 #' unlink(dummypackage, recursive = TRUE)
-inflate <- function(pkg = ".", rmd = file.path("dev", "flat_full.Rmd"),
-                    name = "Get started", check = TRUE, document = TRUE,
+inflate <- function(pkg = ".", flat_file = file.path("dev", "flat_full.Rmd"),
+                    vignette_name = "Get started",
+                    open_vignette = TRUE,
+                    check = TRUE, document = TRUE,
                     overwrite = "ask", ...) {
+
+  if (!is.null(list(...)[["name"]])) {
+    warning(paste0("The `name` argument to `inflate()` is deprecated since {fusen} version 0.2.5,",
+    " and will be removed in a future version.",
+    "\nPlease use `vignette_name = '", list(...)[["name"]],"'` instead.\n"))
+    vignette_name <- list(...)[["name"]]
+  }
+  if (!is.null(list(...)[["rmd"]])) {
+    warning(paste0("The `rmd` argument to `inflate()` is deprecated since {fusen} version 0.2.5,",
+                   " and will be removed in a future version.",
+                   "\nPlease use `flat_file = '", list(...)[["rmd"]],"'` instead.\n"))
+    flat_file <- list(...)[["rmd"]]
+  }
 
   # Save all open files
   if (
@@ -73,7 +91,7 @@ inflate <- function(pkg = ".", rmd = file.path("dev", "flat_full.Rmd"),
   }
 
   pkg <- normalizePath(pkg)
-  rmd <- normalizePath(rmd, mustWork = FALSE)
+  flat_file <- normalizePath(flat_file, mustWork = FALSE)
 
   if (!file.exists(file.path(normalizePath(pkg), "DESCRIPTION"))) {
     stop(
@@ -86,20 +104,20 @@ inflate <- function(pkg = ".", rmd = file.path("dev", "flat_full.Rmd"),
     if (!file.exists(".Rbuildignore")) {
       file.create(".Rbuildignore")
     }
-    # usethis::use_build_ignore(basename(rmd))
+    # usethis::use_build_ignore(basename(flat_file))
     usethis::use_build_ignore(paste0(basename(pkg), ".Rproj"))
     usethis::use_build_ignore(".Rproj.user")
   }
 
-  if (grepl(pkg, rmd, fixed = TRUE)) {
+  if (grepl(pkg, flat_file, fixed = TRUE)) {
     # Rmd already contains pkgpath
-    rmd_path <- rmd
+    flat_file_path <- flat_file
   } else {
-    rmd_path <- file.path(pkg, rmd)
+    flat_file_path <- file.path(pkg, flat_file)
   }
 
-  if (!file.exists(rmd_path)) {
-    stop(rmd, " does not exists, please use fusen::add_flat_template() to create it.")
+  if (!file.exists(flat_file_path)) {
+    stop(flat_file, " does not exists, please use fusen::add_flat_template() to create it.")
   }
 
   # Are you sure ?
@@ -107,8 +125,8 @@ inflate <- function(pkg = ".", rmd = file.path("dev", "flat_full.Rmd"),
     overwrite <- ifelse(isTRUE(overwrite), "yes", "no")
   }
   overwrite <- match.arg(overwrite, choices = c("ask", "yes", "no"))
-  cleaned_name <- asciify_name(name)
-  vignette_path <- file.path(pkg, "vignettes", paste0(cleaned_name, ".Rmd"))
+  cleaned_vignette_name <- asciify_name(vignette_name)
+  vignette_path <- file.path(pkg, "vignettes", paste0(cleaned_vignette_name, ".Rmd"))
   if (file.exists(vignette_path)) {
     if (overwrite == "ask") {
       rm_exist_vignette <- getFromNamespace("can_overwrite", "usethis")(vignette_path)
@@ -131,8 +149,8 @@ inflate <- function(pkg = ".", rmd = file.path("dev", "flat_full.Rmd"),
     roxygen2::roxygenise(pkg)
   }
 
-  parsed_rmd <- parse_rmd(rmd)
-  parsed_tbl <- as_tibble(parsed_rmd)
+  parsed_flat_file <- parse_rmd(flat_file)
+  parsed_tbl <- as_tibble(parsed_flat_file)
 
   parsed_tbl$order <- 1:nrow(parsed_tbl)
   sec_title <- paste(parsed_tbl[["sec_h1"]],
@@ -151,11 +169,11 @@ inflate <- function(pkg = ".", rmd = file.path("dev", "flat_full.Rmd"),
   if (!is.null(fun_code)) {
     create_functions_all(parsed_tbl, fun_code, pkg)
   } else {
-    message("No chunks named 'function-xx' or 'fun-xx' were found in the Rmarkdown file: ", rmd)
+    message("No chunks named 'function-xx' or 'fun-xx' were found in the Rmarkdown file: ", flat_file)
   }
 
   # Create vignette ----
-  create_vignette(parsed_tbl, pkg, name)
+  create_vignette(parsed_tbl, pkg, vignette_name, open_vignette = open_vignette)
 
   # Run attachment
   if (isTRUE(document)) {
@@ -229,67 +247,6 @@ create_functions_all <- function(parsed_tbl, fun_code, pkg) {
   create_tests_files(parsed_tbl, pkg)
 }
 
-#' Parse function code as tibble and get positions
-#' @param x One row out of function parsed tibble
-#' @noRd
-parse_fun <- function(x) { # x <- rmd_fun[3,]
-
-  code <- rmd_get_chunk(x)$code
-  # find function name
-  fun_name <- stringr::str_extract(
-    code[grep("function(\\s*)\\(", code)],
-    "[\\w[.]]*(?=(\\s*)(<-|=)(\\s*)function)"
-  ) %>%
-    gsub(" ", "", .) # remove spaces
-
-  # Clean extra space between #' and @
-  code <- gsub(pattern = "#'\\s*@", "#' @", code)
-
-  # Find start of function
-  first_function_start <- grep("function(\\s*)\\(", code)[1]
-  # Get all #'
-  all_hastags <- grep("^#'", code)
-  if (length(all_hastags) != 0) {
-    last_hastags_above_first_fun <- max(all_hastags[all_hastags < first_function_start])
-  } else {
-    last_hastags_above_first_fun <- NA
-  }
-
-  # Add @noRd if no roxygen doc or no @export before function
-  if (!any(grepl("@export|@noRd", code))) {
-    if (!is.na(last_hastags_above_first_fun)) {
-      code <- c(
-        code[1:last_hastags_above_first_fun],
-        "#' @noRd",
-        code[(last_hastags_above_first_fun + 1):length(code)]
-      )
-    } else if (all(grepl("^\\s*$", code))) {
-      # If all empty
-      code <- character(0)
-    } else if (!is.na(first_function_start)) {
-      # If there is a function inside
-      code <- c("#' @noRd", code)
-    }
-    # otherwise code stays code
-  }
-
-  all_arobase <- grep("^#'\\s*@|function(\\s*)\\(", code)
-  example_pos_start <- grep("^#'\\s*@example", code)[1]
-
-  example_pos_end <- all_arobase[all_arobase > example_pos_start][1] - 1
-  example_pos_end <- ifelse(is.na(example_pos_end),
-    grep("function(\\s*)\\(", code) - 1,
-    example_pos_end
-  )
-
-  tibble::tibble(
-    fun_name = fun_name[1],
-    code = list(code),
-    example_pos_start = example_pos_start,
-    example_pos_end = example_pos_end
-  )
-}
-
 #' Get function names ----
 #' @param parsed_tbl tibble of a parsed Rmd
 #' @importFrom parsermd rmd_get_chunk
@@ -309,128 +266,6 @@ get_functions <- function(parsed_tbl) {
   } else {
     return(NULL)
   }
-}
-
-#' Add function name to parsed_tbl ----
-#' @param parsed_tbl tibble of a parsed Rmd
-#' @param fun_names Names of functions in Rmd
-#' @importFrom stats na.omit
-#' @noRd
-add_fun_to_parsed <- function(parsed_tbl, fun_names) {
-  which_parsed_fun <- which(!is.na(parsed_tbl$label) &
-    grepl(regex_functions, parsed_tbl$label))
-
-  parsed_tbl$fun_name <- NA_character_
-  # Function name
-  parsed_tbl[["fun_name"]][which_parsed_fun] <- fun_names
-
-  pkg_filled <- lapply(na.omit(unique(parsed_tbl[["sec_title"]])), function(x) {
-    group <- which(parsed_tbl[["sec_title"]] == x)
-    parsed_tbl[group, ] <- tidyr::fill(parsed_tbl[group, ], fun_name)
-  }) %>%
-    do.call("rbind", .)
-  parsed_tbl[["fun_name"]][pkg_filled[["order"]]] <- pkg_filled[["fun_name"]]
-  parsed_tbl
-}
-
-#' Add examples in function code
-#' @param parsed_tbl tibble of a parsed Rmd
-#' @param fun_code R code of functions in Rmd as character
-#' @noRd
-add_fun_code_examples <- function(parsed_tbl, fun_code) {
-  # fun_code <- fun_code[!is.na(fun_code[["fun_name"]]), ]
-  #  Example already in skeleton
-  fun_code$example_in <- apply(fun_code, 1, function(x) {
-    if (is.na(x[["fun_name"]])) {
-      list(character(0))
-    } else if (!is.na(x[["example_pos_start"]]) && length(x[["example_pos_start"]]) == 1) {
-      list(x[["code"]][x[["example_pos_start"]]:x[["example_pos_end"]]])
-    } else {
-      list("#' @examples")
-    }
-  }) %>% lapply(., function(x) x[[1]])
-
-  # Example in separate chunk
-  which_parsed_ex <- which(!is.na(parsed_tbl$label) &
-    grepl(regex_example, parsed_tbl$label))
-  rmd_ex <- parsed_tbl[which_parsed_ex, ]
-  # No function, no example to add
-  rmd_ex <- rmd_ex[!is.na(rmd_ex[["fun_name"]]), ]
-
-
-  if (nrow(rmd_ex) != 0) {
-    example_code <- lapply(
-      seq_len(nrow(rmd_ex)),
-      function(x) {
-        tibble::tibble(
-          fun_name = rmd_ex[x, ][["fun_name"]],
-          example_chunk = list(paste("#'", rmd_get_chunk(rmd_ex[x, ])$code))
-        )
-      }
-    ) %>% do.call("rbind", .)
-    # Add to function tibble
-    fun_code <- merge(fun_code, example_code, by = "fun_name", all.x = TRUE) %>%
-      tibble::as_tibble()
-    fun_code[["example"]] <- lapply(seq_len(nrow(fun_code)), function(x) {
-      example <- stats::na.omit(unlist(c(
-        fun_code[["example_in"]][x],
-        fun_code[["example_chunk"]][x]
-      )))
-    })
-  } else {
-    fun_code[["example"]] <- fun_code[["example_in"]]
-  }
-
-  # Remove if example is empty
-  fun_code[["example"]] <- lapply(fun_code[["example"]], function(example) {
-    # example <- fun_code[["example"]][[1]]
-    if (length(example) == 0) {
-      return(NA)
-    } else if (length(example) == 1 && is.na(example)) {
-      return(NA)
-    } else if (length(example) == 1 && example == "#' @examples") {
-      return(NA)
-    } else if (length(example) > 1 & all(grepl("^#'\\s+$", example[-1]))) {
-      return(NA)
-    } else {
-      return(example)
-    }
-  })
-
-  # Add to function code
-  fun_code[["code_example"]] <- lapply(seq_len(nrow(fun_code)), function(x) {
-
-    fun_code_x <- fun_code[x, ]
-    if (is.na(fun_code_x[["fun_name"]])) {
-      return(
-        unlist(fun_code_x[["code"]]))
-    }
-
-    end_skeleton <- ifelse(is.na(fun_code_x[["example_pos_start"]]),
-      fun_code_x[["example_pos_end"]],
-      fun_code_x[["example_pos_start"]] - 1
-    )
-
-    all_fun_code <- stats::na.omit(c(
-      # begin
-      if (!is.na(end_skeleton)) {
-        unlist(fun_code_x[["code"]])[1:end_skeleton]
-      },
-      # examples
-      unlist(fun_code_x[["example"]]),
-      # end
-      unlist(fun_code_x[["code"]])[
-        (fun_code_x[["example_pos_end"]] + 1):length(unlist(fun_code_x[["code"]]))
-      ]
-    ))
-  })
-
-  # Clean double #' due to dontrun
-  fun_code[["code_example"]] <- lapply(fun_code[["code_example"]], function(example) {
-    gsub("#' #' ", "#' ", example)
-  })
-
-  return(fun_code)
 }
 
 #' create R file with code content and fun name
@@ -467,11 +302,11 @@ create_r_files <- function(fun_code, pkg) {
     if (file.exists(r_file)) {
       cli::cli_alert_warning(paste(basename(r_file), "has been overwritten"))
     }
-    cat(
+    lines <- c(
       "# Generated by fusen: do not edit by hand\n",
-      enc2utf8(unlist(fun_code[x, ][["code_example"]])),
-      file = r_file, sep = "\n"
+      unlist(fun_code[x, ][["code_example"]])
     )
+    write_utf8(path = r_file, lines = lines)
     r_file
   })
 }
@@ -479,7 +314,7 @@ create_r_files <- function(fun_code, pkg) {
 #' Check if there are unit tests ----
 #' @param parsed_tbl tibble of a parsed Rmd
 #' @param pkg Path to package
-#' @importFrom parsermd rmd_get_chunk
+#' @importFrom parsermd rmd_node_code
 #' @noRd
 create_tests_files <- function(parsed_tbl, pkg) {
   rmd_test <- parsed_tbl[!is.na(parsed_tbl$label) &
@@ -506,7 +341,10 @@ create_tests_files <- function(parsed_tbl, pkg) {
     }
 
     parse_test <- function(x) { # x <- rmd_test[1,]
-      code <- rmd_get_chunk(x)$code
+      # code <- rmd_get_chunk(x)$code
+      code <- unlist(rmd_node_code(x$ast))
+      # parsermd::rmd_select(dev_parse, "description")[[1]] %>%
+      #   parsermd::rmd_node_code()
 
       # create R file with code content and fun name
       fun_name <- x[["fun_name"]]
@@ -518,10 +356,10 @@ create_tests_files <- function(parsed_tbl, pkg) {
       if (file.exists(test_file)) {
         cli::cli_alert_warning(paste(basename(test_file), "has been overwritten"))
       }
-      cat(
+      lines <- c(
         "# Generated by fusen: do not edit by hand\n",
-        enc2utf8(code), file = test_file, sep = "\n")
-
+        code)
+      write_utf8(path = test_file, lines = lines)
       fun_name
     }
 
@@ -550,9 +388,10 @@ create_tests_files <- function(parsed_tbl, pkg) {
 #' Create vignette
 #' @param parsed_tbl tibble of a parsed Rmd
 #' @param pkg Path to package
-#' @param name Name of the resulting vignette
+#' @param vignette_name Name of the resulting vignette
+#' @param open_vignette Logical. Whether to open vignette file
 #' @noRd
-create_vignette <- function(parsed_tbl, pkg, name) {
+create_vignette <- function(parsed_tbl, pkg, vignette_name, open_vignette = TRUE) {
   old_proj <- usethis::proj_get()
 
   if (normalizePath(old_proj) != normalizePath(pkg)) {
@@ -585,7 +424,7 @@ create_vignette <- function(parsed_tbl, pkg, name) {
   # vignette_tbl[["label"]][grepl("unnamed", vignette_tbl[["label"]])] <-
   #   gsub("unnamed-", "parsermd-", vignette_tbl[["label"]][grepl("unnamed", vignette_tbl[["label"]])])
   #   is.na(vignette_tbl[["label"]]) & vignette_tbl[["type"]] == "rmd_chunk",
-  #                                   gsub("[.]+", "-", make.names(name)),
+  #                                   gsub("[.]+", "-", make.names(vignette_name)),
   #                                   vignette_tbl[["label"]])
   #
   # vignette_tbl[["label"]] <- make.unique(vignette_tbl[["label"]], sep = "-")
@@ -605,32 +444,37 @@ create_vignette <- function(parsed_tbl, pkg, name) {
   vignette_tbl[["ast"]] <- lapply(vignette_tbl[["ast"]], fix_unnamed_chunks)
   class(vignette_tbl[["ast"]]) <- ast_class
 
-  cleaned_name <- asciify_name(name)
+  # File to save
+  cleaned_vignette_name <- asciify_name(vignette_name)
+  vignette_file <- file.path("vignettes", paste0(cleaned_vignette_name, ".Rmd"))
 
-  usethis::use_vignette(name = cleaned_name, title = name)
-  vignette_file <- file.path("vignettes", paste0(cleaned_name, ".Rmd"))
-  if (!file.exists(vignette_file)) {
-    stop(
-      "Vignette could not be filled because of naming problem.",
-      "Have you used some special characters in `name`?"
-    )
-  }
+  # usethis::use_vignette(name = cleaned_vignette_name, title = vignette_name)
+  # if (!file.exists(vignette_file)) {
+  #   stop(
+  #     "Vignette could not be filled because of naming problem.",
+  #     "Have you used some special characters in `vignette_name`?"
+  #   )
+  # }
+
+  # Vignette head
+  head <- create_vignette_head(pkg = pkg, vignette_name = vignette_name)
 
   # Write vignette
   if (nrow(vignette_tbl) == 0) {
-    cat(
-      "",
-      "<!-- This vignette is generated by fusen: do not edit by hand -->\n",
-      sep = "\n", append = TRUE,
-      file = vignette_file
+    lines <- c(head,
+               "",
+               "<!-- This vignette is generated by fusen: do not edit by hand -->\n"
     )
   } else {
-    cat(
-      "",
-      "<!-- This vignette is generated by fusen: do not edit by hand -->\n",
-      enc2utf8(parsermd::as_document(vignette_tbl)),
-      sep = "\n", append = TRUE,
-      file = vignette_file
-    )
+    lines <- c(
+        head,
+        "",
+        "<!-- This vignette is generated by fusen: do not edit by hand -->\n",
+        parsermd::as_document(vignette_tbl)
+      )
   }
+  write_utf8(path = vignette_file, lines = lines)
+
+  if (isTRUE(open_vignette) & interactive()) {usethis::edit_file(vignette_file)}
 }
+